Included while loop on process termination.

+++ src/website/run_dash_in_background.sh
@@ -23,7 +23,6 @@ port_is_occupied() {
 
 assert_port_is_occupied() {
   local local_project_port="$1"
-
   if [[ "$(port_is_occupied "$local_project_port")" != "FOUND" ]]; then
     echo "Error, port:$local_project_port is not occupied:$(sudo lsof -i:"$local_project_port")"
     exit 6
@@ -38,17 +37,21 @@ assert_port_is_free() {
     exit 6
   fi
 }
-terminate_process_on_port() {
+
+terminate_processes_on_port() {
   local local_project_port="$1"
+  # TODO: include counter and limit on nr of checks.
+  while true; do
+    if [[ "$(port_is_occupied "$local_project_port")" != "NOTFOUND" ]]; then
+      local pids
+      pids="$(sudo lsof -t -i:"$local_project_port")"
+      local first_pid
+      first_pid=${pids%$'\n'*}
+      echo "For port:$local_project_port, killing process:$first_pid"
+      sudo kill -9 "$(lsof -t -i:"$first_pid")"
+    else
+      assert_port_is_free "$local_project_port"
+    fi
+  done
 
-  local pids
-  pids="$(sudo lsof -t -i:"$local_project_port")"
-  read -p "pids=$pids-ENDWITHOUTSPACE"
-  local first_pid
-  first_pid=${pids%$'\n'*}
-  read -p "first_pid=$first_pid-ENDWITHOUTSPACE"
-
-  # sudo kill -9 "$(lsof -t -i:"$local_project_port")"
-  sudo kill -9 "$(lsof -t -i:"$first_pid")"
-  assert_port_is_free "$local_project_port"
 }
